+++ io-takamaka-code/src/main/java/io/takamaka/code/governance/Manifest.java
@@ -55,12 +55,12 @@ public final class Manifest<V extends Validator> extends ExternallyOwnedAccount 
 	 * The maximal length of the error message kept in the store of the node.
 	 * Beyond this threshold, the message gets truncated.
 	 */
-	private final int maxErrorLength;
+	private final long maxErrorLength;
 
 	/**
 	 * The maximal number of dependencies in the classpath of a transaction.
 	 */
-	private final int maxDependencies;
+	private final long maxDependencies;
 
 	/**
 	 * The maximal cumulative size (in bytes) of the instrumented jars of the dependencies of a transaction.
@@ -144,7 +144,7 @@ public final class Manifest<V extends Validator> extends ExternallyOwnedAccount 
 	 * @param builderOfGasStation the builder of the gas station of the node having the manifest
 	 * @throws RequirementViolationException if any parameter is null or any builder yields null or the maximal error length is negative
 	 */
-	public Manifest(String genesisTime, String chainId, int maxErrorLength, int maxDependencies, long maxCumulativeSizeOfDependencies, boolean allowsSelfCharged,
+	public Manifest(String genesisTime, String chainId, long maxErrorLength, long maxDependencies, long maxCumulativeSizeOfDependencies, boolean allowsSelfCharged,
 			boolean allowsFaucet, boolean allowsMintBurnFromGamete, boolean skipsVerification, String signature, Gamete gamete, long verificationVersion,
 			Function<Manifest<V>, Validators<V>> builderOfValidators, Function<Manifest<V>, GasStation<V>> builderOfGasStation) {
 
@@ -204,7 +204,7 @@ public final class Manifest<V extends Validator> extends ExternallyOwnedAccount 
 	 * 
 	 * @return the length
 	 */
-	public final @View int getMaxErrorLength() {
+	public final @View long getMaxErrorLength() {
 		return maxErrorLength;
 	}
 
@@ -214,7 +214,7 @@ public final class Manifest<V extends Validator> extends ExternallyOwnedAccount 
 	 * 
 	 * @return the maximal number of dependencies
 	 */
-	public final @View int getMaxDependencies() {
+	public final @View long getMaxDependencies() {
 		return maxDependencies;
 	}
